bulk changes
