+++ app/components/gcse_row_content.html.erb
