+++ app/components/gcse_row_content.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+class GcseRowContent < GcseRequirements
+end
