--- app/components/gcse_row_content_component/view.rb
@@ -1,6 +1,0 @@
-# frozen_string_literal: true
-
-module GcseRowContentComponent
-  class View < GcseRequirements
-  end
-end
