+++ app/components/paginated_filter.html.erb
@@ -8,6 +8,6 @@
 
     <%= content %>
 
-    <%= render Paginator::View.new(scope: collection) %>
+    <%= render Paginator.new(scope: collection) %>
   </div>
 </div>
