+++ app/components/persona.html.erb
