+++ app/components/phase_banner.html.erb
