+++ app/components/phase_banner.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-class PhaseBanner::View < ViewComponent::Base
+class PhaseBanner < ViewComponent::Base
   def initialize(no_border: false)
     super
     @no_border = no_border
