+++ app/components/support_title_bar.html.erb
