+++ app/components/tab_navigation.html.erb
