+++ app/components/tab_navigation.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+class TabNavigation < ApplicationComponent
+  attr_reader :items
+
+  def initialize(items:, classes: [], html_attributes: {})
+    super(classes:, html_attributes:)
+    @items = items
+  end
+end
