+++ app/components/title_bar.html.erb
