+++ app/views/publish/courses/_description_content.html.erb
@@ -137,7 +137,7 @@
       summary_list,
       :course,
       "GCSEs",
-      (render GcseRowContentComponent::View.new(course:, errors: @errors)),
+      (render GcseRowContent.new(course:, errors: @errors)),
       %w[accept_pending_gcse accept_gcse_equivalency accept_english_gcse_equivalency accept_maths_gcse_equivalency accept_science_gcse_equivalency additional_gcse_equivalencies],
       truncate_value: false,
       action_path: course.gcse_section_complete? ? gcses_pending_or_equivalency_tests_publish_provider_recruitment_cycle_course_path(@provider.provider_code, @provider.recruitment_cycle_year, course.course_code) : nil,
