+++ app/views/support/courses/index.html.erb
@@ -22,4 +22,4 @@
   </tbody>
 </table>
 
-<%= render Paginator::View.new(scope: @courses) %>
+<%= render Paginator.new(scope: @courses) %>
