+++ app/views/support/locations/index.html.erb
@@ -36,4 +36,4 @@
   </tbody>
 </table>
 
-<%= render Paginator::View.new(scope: @sites) %>
+<%= render Paginator.new(scope: @sites) %>
