+++ app/views/support/providers/users/index.html.erb
@@ -1,3 +1,3 @@
 <%= link_to t("components.page_titles.support.providers.users.new"), new_support_recruitment_cycle_provider_user_path(@provider.recruitment_cycle_year, @provider), class: "govuk-!-margin-top-4 govuk-button" %>
 <%= render "provider_user_list", users: @users, provider: @provider %>
-<%= render Paginator::View.new(scope: @users) %>
+<%= render Paginator.new(scope: @users) %>
