+++ app/views/support/users/providers.html.erb
@@ -2,4 +2,4 @@
 
 <%= render "support/users/user_providers", providers: @providers %>
 
-<%= render Paginator::View.new(scope: @providers) %>
+<%= render Paginator.new(scope: @providers) %>
