--- spec/components/support_title_bar/view_preview.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-module SupportTitleBar
-  class ViewPreview < ViewComponent::Preview
-    def default
-      render(View.new)
-    end
-  end
-end
