+++ spec/components/support_title_bar_preview.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class SupportTitleBarPreview < ViewComponent::Preview
+  def default
+    render(SupportTitleBar.new)
+  end
+end
